training/training_requirements.txt
tensorflow-gpu==2.10.0
numpy==1.23.5
tqdm==4.66.4
# pickle-mixin==1.0.2
# flask==3.0.3
# streamlit==1.19.0
# protobuf==3.20.3
scikit-learn==1.2.2
ipykernel
matplotlib
seaborn
Pillow
